docs/index.rst: Remove line blocks
Avoid errors like:

  .../docs/index.rst:26: WARNING: Line block ends without a blank line.

by dropping the line blocks [1] added in 323ec78d (Update
documentation of each parameter, 2014-07-29).  I don't think we want
to manually wrap the line endings here, and even if we did the syntax
was invalid (continuation lines need to begin with a space in place of
the vertical bar [1]).

Since I was touching that section of the docs, I also replaced the
manually-escaped 'CORS\_' and similar with inline literals [2].
Besides allowing us to use the more-easily-grepped 'CORS_', it gives
us nicer styling.

[1]: http://docutils.sourceforge.net/docs/ref/rst/restructuredtext.html#line-blocks
[2]: http://docutils.sourceforge.net/docs/ref/rst/restructuredtext.html#inline-literals