Fix regex sorting issue from #391 (#392)
* Rename "helper_tests.py" to "test_helpers.py".

This causes it to get picked up by pytest, which makes clear the failure
that I will fix in a followup commit.

* Sort paths longest to shortest.

Previously, path sort length was inverted, with short paths before long
ones, this led to a path like `/*` being handled *before* a path like
`/foo/bar/baz`, which is exactly what we didn't want. This was tested,
but the test was misnamed and so wasn't picked up by pytest.

---------

Co-authored-by: joshuamorton <joshuamorton@google.com>